requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-RESTful==0.3.9
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.9
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.17.0
